backtest/requirements.txt
flask
gunicorn
python-dotenv
backtrader
yfinance
matplotlib
flasgger